+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## 8.2.1 (git master)
+* Update the .net extension method ToIEnumerable to work with java.lang.Iterable rather than java.util.Collection
 
 ## 8.2.0 (09/09/2020)
 * All readers, including the UniversalProjectReader, now support a readAll method. If a file or database contains more than one project the readAll method can be used to retrieve them all in one operation. If the file format doesn't support multiple schedules, readAll will just return a single schedule.